+++ changelog.md
@@ -16,6 +16,7 @@ __[maintenance]__
 * Update certificate information for Eclipse Temurin JDK 21 LTS installer.
 * Update certificate information for Eclipse Temurin JRE 21 LTS installer.
 * Update certificate information for Eclipse Temurin JDK 25 LTS installer.
+* Update certificate information for Eclipse Temurin JRE 25 LTS installer.
 
 ## Version 2026.07.18.0